Main

Documentation

main :: IO ()

Select the main function of your choice: main1, main2, ...

Types and data constructors

type Position = V2 Float

type Velocity = V2 Float

type Index = Int

type Force = V2 Float

type Forces = [Force]

type TimeStep = Float

data Particle

Constructors

Particle

Fields

idx :: Index

pos :: Position

vel :: Velocity

type Particles = [Particle]

type Model = Particles

Constant functions (throughout all main)

dotSize :: Float

windowDisplay :: Display

simulationRate :: Int

toPixels :: Float -> Float

Simulations

First simulation"

main1 :: IO ()

The initial model consists of one particle, which will be drawn and updated according to newton.

This results in a particle moving from the center of the screen to the right:

Helper functions

drawParticles :: Particles -> [Picture]

Convert a list of Particle (Particles) into a list of Picture

drawParticle :: Particle -> Picture

Visualize a single particle as blue dot

newton :: Float -> Particles -> Particles

Update velocity of one particle assuming no acceleration