Main

Documentation

main :: IO ()

Select the main function of your choice

mainNewton

mainNewtonBounce

mainVerlet

mainVerletSquare

mainVerletRandom

to perform the according simulation

Types and data constructors

type Position = V2 Float

Represents the spatial vector (Position) \(\vec{r}\) as a two-dimensional vector

type Velocity = V2 Float

Represents the velocity vector \(\vec{v}\) as a two-dimensional vector

type Index = Int

Index of a Particle to make it distinguishable from other Particles

Note: Position and Velocity do not make a Particle distinguishable and even if, floating point errors would make this comparison invalid

type Force = V2 Float

Represents the force vector \(\vec{F}\) as a two-dimensional vector

type TimeStep = Float

Represents the time step in the simulation \( (\Delta t) \)

data Particle

Representation of a Particle. All we need is a Index (comparison) and a Position and Velocity. The mass of the particle m is assumed to be the same for all Particles

Constructors

Particle

Fields

idx :: Index

pos :: Position

vel :: Velocity

type Model = [Particle]

The Model consists of a set/list of Particle

Constant functions (throughout all main)

dotSize :: Float

Size of the Particles

windowDisplay :: Display

Window display config

Show a 800x800 window at the window coordinate (200, 800) with the title "MD in Haskell"

simulationRate :: Int

Number of frames per second

toPixels :: Float -> Float

Transform a value to the scale of Pixels

aLength :: Float

Length in the x dimension (width)

bLength :: Float

Length in the y dimension (height)

Simulations

First simulation

mainNewton :: IO ()

The initial model consists of one Particle, which will be drawn and updated according to newton.

This results in a Particle moving from the center of the screen to the right:

Helper functions

drawParticle :: Particle -> Picture

Visualize a single Particle as blue dot by transforming the Position of the Particle into a Picture, which can be rendered

newton :: TimeStep -> [Particle] -> [Particle]

Update velocity of one Particle assuming no acceleration

Second simulation

mainNewtonBounce :: IO ()

The initial model consists of one Particle, which will be drawn and updated according to newtonBounce.

This results in a Particle moving from the center of the screen to the right and will bounce of the wall:

Additional helper functions

drawWalls :: Picture

A Picture of the wall surrounding the Particle ("simulation box")

newtonBounce :: Float -> [Particle] -> [Particle]

Updated newton function, which incorporates bouncing of the wall (see boundaryCondition)

boundaryCondition :: Particle -> V2 Float

Transformation vector, which is used to change the direction of a Particle, when it is close (current Position + radius of the particle) to a wall in x, y or (x and y) direction.

Third simulation

mainVerlet :: IO ()

First simulation using the Velocity Verlet algorithm and the Lennard-Jones potential of two Particles. They attract and repulse each other.

Additional helper functions

verletStep :: TimeStep -> Model -> Model

Implementation of the Velocity Verlet algorithm (without the half-step)

m :: Float

Mass of all our Particles

epsilon :: Float

The \(\epsilon\) value of the Lennard-Jones potential

sigma :: Float

The \(\sigma\) value of the Lennard-Jones potential

sigma6 :: Float

Helper function for

\(\sigma^6\)

sigma12 :: Float

Helper function for

\(\sigma^{12}\)

calcForces :: [Particle] -> [Force]

Calculate all Forces on all Particles employing calcForceAcc

calcForceAcc :: [Particle] -> [Particle] -> [Force]

Helper function for calculating all Forces on all Particles

calcForceOnOne :: Particle -> [Particle] -> [Force]

Calculate all Forces on one Particle

calcForceBetween :: Particle -> Particle -> Force

Calculates the Force between two Particle as long as they are not the same (compare Index)

repulsion :: Position -> Position -> Force

Repulsion term of the Lennard-Jones potential

attraction :: Position -> Position -> Force

Attraction term of the Lennard-Jones potential

updatePosition :: TimeStep -> Particle -> Force -> Particle

Update the position/velocity of one Particle

updateVelocity :: TimeStep -> Particle -> Force -> Particle

Update the position/velocity of one Particle

updatePositions :: TimeStep -> [Particle] -> [Force] -> [Particle]

Apply updatePosition/updateVelocity to all (list/set) Particles

updateVelocities :: TimeStep -> [Particle] -> [Force] -> [Particle]

Apply updatePosition/updateVelocity to all (list/set) Particles

Fourth simulation

mainVerletSquare :: IO ()

Same as mainVerlet but with a square lattice of \(4 \times 4\) Particles

Additional helper functions

squareLatticeModel :: Int -> [Particle]

Function to construct a square lattice of dimension \(n \times n\)

squareLattice :: Int -> Int -> [Position]

Constructs the square lattice (squareLatticeModel) row by row. Each row has the same distance from the next row

latticeRow :: Int -> Int -> Float -> [Position]

Constructs a even-distanced row of Particles

Fourth simulation

mainVerletRandom :: IO ()

Same as mainVerlet but with 16 random generated Particles

Additional helper functions

modelRandom :: RandomGen g => Int -> g -> [Particle]

Generates \(n\) random Particles (random Position and Velocity).

randomVel :: RandomGen g => Int -> g -> [Velocity]

Generates \(n\) random Velocity values

Range: \(v \in [-0.2,0.2]\)

randomPos :: RandomGen g => Int -> g -> [Position]

Generates \(n\) random Positions

genPos :: RandomGen g => g -> (Position, g)

Generates \(1\) random Position

Range: All values inside the simulation box (drawWalls)